When JSON is used, let (packed) sequences be arrays.
A PackedSeq was previously represented as a String in JSON. Using an
Array makes manipulation of the output using programs such as jq
easier. The only downside is that JSON cannot represent all
Rationals. E.g. representing 1/3 in JSON results in loss of
precision. I opted to throw an error in such cases.